\name{pre}
\alias{pre}
\title{Predict the best OC threshold (T) via OC of prents }
\usage{
pre(fpoc,ioc,hir)
}
\description{

fpoc: Kernel oil content of female parent

ioc: Kernel oil content of inducers

hir: Haploid induct rate

}
\examples{
## Estimate the best  threshold (T) via OC of prents

## pre(fpoc,ioc,hir)

pre(2.8,8,0.15)

}
